+++ components/reservations/BasicButton.tsx
@@ -9,7 +9,7 @@ export default function BasicButton({
   return (
     <button
       type={type}
-      className={`text-xs font-normal text-neutral-700 rounded-sm border border-neutral-200 ${className}`}
+      className={`bg-white text-neutral-700 disabled:text-neutral-300 rounded-sm border border-neutral-200 ${className}`}
       {...props}
     >
       {children}
