+++ components/reservations/modals/useAddReservation.ts
@@ -12,73 +12,71 @@ import { infoToast, errorToast } from '@/utils/toast';
 
 import getOptimalEndTime from './getOptimalEndTime';
 
+export type SetReservationBody = <T extends keyof ReservationPostBody>(
+  key: T,
+  value: ReservationPostBody[T],
+) => void;
+
 export default function useAddReservation(roomId: number) {
+  const [body, _setBody] = useState<ReservationPostBody>(() => getDefaultBodyValue(roomId));
   const [privacyChecked, togglePrivacyChecked] = useReducer((x) => !x, false);
-  const [body, _setBody] = useState<ReservationPostBody>(getDefaultBodyValue(roomId));
+  const canSubmit = checkSubmit(privacyChecked, body);
 
   const handleSubmit: FormEventHandler<HTMLFormElement> = async (e) => {
     e.preventDefault();
-
-    const canSubmit =
-      privacyChecked &&
-      body.title !== '' &&
-      body.contactEmail !== '' &&
-      body.professor !== '' &&
-      body.purpose !== '';
 
     if (!canSubmit) {
       infoToast('모든 필수 정보를 입력해주세요');
       return;
     }
+
     try {
       await postReservation(body);
       // TODO: revalidate같은거 써서 좀 더 예쁘게
       refreshPage();
     } catch (e) {
-      if (e instanceof NetworkError) {
-        if (e.statusCode === 409) {
-          errorToast('해당 위치에 이미 예약이 존재합니다.');
-        } else {
-          errorToast(e.message);
-        }
-      } else if (e instanceof Error) {
-        errorToast(e.message);
-      } else {
-        errorToast('알 수 없는 에러');
-      }
+      handleError(e);
     }
   };
 
-  const setBody = <T extends keyof ReservationPostBody>(key: T, value: ReservationPostBody[T]) =>
+  const setBody: SetReservationBody = (key, value) =>
     _setBody((body) => ({ ...body, [key]: value }));
 
-  // 날짜 선택 모달용
+  // 날짜 선택시 시작 시간과 끝나는 시간을 업데이트
   const setDate = (date: Date) => {
-    let newStartTime = new Date(body.startTime);
-    newStartTime.setFullYear(date.getFullYear());
-    newStartTime.setMonth(date.getMonth());
-    newStartTime.setDate(date.getDate());
+    let newStart = new Date(body.startTime);
+    newStart.setFullYear(date.getFullYear());
+    newStart.setMonth(date.getMonth());
+    newStart.setDate(date.getDate());
 
     // 오늘을 선택헀으면 예약 불가 시간 필터링
-    if (isSameDay(newStartTime, new Date())) {
-      newStartTime = convertToFastestStartTime(newStartTime);
+    if (isSameDay(newStart, new Date())) {
+      newStart = getEarliestStartTimeFrom(newStart);
     }
 
-    const endTime = getOptimalEndTime(
-      new Date(body.startTime),
-      new Date(body.endTime),
-      newStartTime,
-    );
+    const endTime = getOptimalEndTime({
+      prevStart: new Date(body.startTime),
+      prevEnd: new Date(body.endTime),
+      newStart,
+    });
 
-    setBody('startTime', newStartTime.toISOString());
+    setBody('startTime', newStart.toISOString());
     setBody('endTime', endTime.toISOString());
   };
 
-  return { handleSubmit, body, setBody, setDate, privacyChecked, togglePrivacyChecked };
+  return {
+    handleSubmit,
+    body,
+    setBody,
+    setDate,
+    privacyChecked,
+    togglePrivacyChecked,
+    canSubmit,
+  };
 }
 
 const getDefaultBodyValue = (roomId: number): ReservationPostBody => {
-  const startTime = convertToFastestStartTime(new Date());
+  const startTime = getEarliestStartTimeFrom(new Date());
 
   const endTime = new Date(startTime);
   endTime.setTime(endTime.getTime() + 30 * 60 * 1000);
@@ -96,8 +94,32 @@ const getDefaultBodyValue = (roomId: number): ReservationPostBody => {
   };
 };
 
+const checkSubmit = (privacyChecked: boolean, body: ReservationPostBody) => {
+  return (
+    privacyChecked &&
+    body.title !== '' &&
+    body.contactEmail !== '' &&
+    body.professor !== '' &&
+    body.purpose !== ''
+  );
+};
+
+const handleError = (e: unknown) => {
+  if (e instanceof NetworkError) {
+    if (e.statusCode === 409) {
+      errorToast('해당 위치에 이미 예약이 존재합니다.');
+    } else {
+      errorToast(e.message);
+    }
+  } else if (e instanceof Error) {
+    errorToast(e.message);
+  } else {
+    errorToast('알 수 없는 에러');
+  }
+};
+
 // date 이후 시점 중 가장 빠른 유효 startTime을 반환
-const convertToFastestStartTime = (date: Date) => {
+const getEarliestStartTimeFrom = (date: Date) => {
   let startTime = new Date(date);
 
   // 과거면 현재로 변경
